requirements.txt
streamlit
pandas
openpyxl
Pillow
easyocr
opencv-python-headless
